Update update-kolmafia.pl
I updated the code to only require the symlink if running on windows. It didn't throw any errors on linux but unfortunately I don't have a way to test it.

Also I was getting the error of: "Use of uninitialized value in string ne at update-kolmafia.pl line 42." this code removed the error for creating the symlink at the bottom (starting at line 59).